requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-dotenv==1.2.1

# ---------- Auth & Security (ADD THESE) ----------
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.10.1
email-validator==2.2.0


# ---------- AI / Agent Stack ----------
langchain==1.1.3
langchain-core==1.2.0
langchain-groq==1.1.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
groq==0.37.1

# ---------- Data & Infra ----------
pymongo==4.15.5
redis==7.1.0

# ---------- HTTP / Utils ----------
requests==2.32.5
httpx==0.28.1
orjson==3.11.5
PyYAML==6.0.3
